pyproject.toml
[tool.poetry]
name = "confluence-poster"
version = "0.1.1"
description = "Project to allow automatic conversion of documents from Markdown or HTML into Confluence pages."
authors = ["Stuart <stuart@email>"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.26.0"
requests-oauthlib = "^1.3.0"
atlassian-python-api = "^3.13.0"
Markdown = "^3.3.4"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
